requirements.txt
# AI Data Lakehouse - Python Dependencies

# Storage & Data Processing
boto3>=1.34.0
pandas>=2.1.0
pyarrow>=14.0.0

# RavenDB Client
ravendb>=5.2.0

# Vector Store
pymilvus>=2.3.0

# Embeddings & ML
sentence-transformers>=2.2.0
torch>=2.1.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0

# HDF5/NeXus Scientific Data
h5py>=3.10.0

# Tiled Data Server (for serving scientific data)
# See: https://blueskyproject.io/tiled/
tiled[all]>=0.1.0a120
s3fs>=2024.2.0

# Jupyter (Optional - for demo notebook)
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0
